Port TestInconsistentResult to serial test package
