+++ tests/e2e/e2e_test.go
@@ -76,128 +76,6 @@ func TestE2E(t *testing.T) {
 		//	},
 		//},
 		testExecution{
-			Name:       "TestInconsistentResult",
-			IsParallel: false,
-			TestFn: func(t *testing.T, f *framework.Framework, ctx *framework.Context, namespace string) error {
-				suiteName := "test-inconsistent"
-				workerScanName := fmt.Sprintf("%s-workers-scan", suiteName)
-				selectWorkers := map[string]string{
-					"node-role.kubernetes.io/worker": "",
-				}
-
-				workersComplianceSuite := &compv1alpha1.ComplianceSuite{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      suiteName,
-						Namespace: namespace,
-					},
-					Spec: compv1alpha1.ComplianceSuiteSpec{
-						ComplianceSuiteSettings: compv1alpha1.ComplianceSuiteSettings{
-							AutoApplyRemediations: false,
-						},
-						Scans: []compv1alpha1.ComplianceScanSpecWrapper{
-							{
-								ComplianceScanSpec: compv1alpha1.ComplianceScanSpec{
-									ContentImage: contentImagePath,
-									Profile:      "xccdf_org.ssgproject.content_profile_moderate",
-									Rule:         "xccdf_org.ssgproject.content_rule_no_direct_root_logins",
-									Content:      rhcosContentFile,
-									NodeSelector: selectWorkers,
-									ComplianceScanSettings: compv1alpha1.ComplianceScanSettings{
-										Debug: true,
-									},
-								},
-								Name: workerScanName,
-							},
-						},
-					},
-				}
-
-				workerNodes := getNodesWithSelectorOrFail(t, f, selectWorkers)
-				pod, err := createAndRemoveEtcSecurettyOnNode(t, f, namespace, "create-etc-securetty", workerNodes[0].Labels["kubernetes.io/hostname"])
-				if err != nil {
-					return err
-				}
-
-				err = f.Client.Create(goctx.TODO(), workersComplianceSuite, getCleanupOpts(ctx))
-				if err != nil {
-					return err
-				}
-
-				// Ensure that all the scans in the suite have finished and are marked as Done
-				err = waitForSuiteScansStatus(t, f, namespace, suiteName, compv1alpha1.PhaseDone, compv1alpha1.ResultInconsistent)
-				if err != nil {
-					E2EErrorf(t, "Got an unexpected status")
-				}
-
-				if err := f.KubeClient.CoreV1().Pods(namespace).Delete(goctx.TODO(), pod.Name, metav1.DeleteOptions{}); err != nil {
-					return err
-				}
-
-				// The check for the no-direct-root-logins rule should be inconsistent
-				var rootLoginCheck compv1alpha1.ComplianceCheckResult
-				rootLoginCheckName := fmt.Sprintf("%s-no-direct-root-logins", workerScanName)
-
-				err = f.Client.Get(goctx.TODO(), types.NamespacedName{Name: rootLoginCheckName, Namespace: namespace}, &rootLoginCheck)
-				if err != nil {
-					return err
-				}
-
-				if rootLoginCheck.Status != compv1alpha1.CheckResultInconsistent {
-					return fmt.Errorf("expected the %s result to be inconsistent, the check result was %s", rootLoginCheckName, rootLoginCheck.Status)
-				}
-
-				var expectedInconsistentSource string
-
-				if len(workerNodes) >= 3 {
-					// The annotations should list the node that had a different result
-					expectedInconsistentSource = workerNodes[0].Name + ":" + string(compv1alpha1.CheckResultPass)
-					inconsistentSources := rootLoginCheck.Annotations[compv1alpha1.ComplianceCheckResultInconsistentSourceAnnotation]
-					if inconsistentSources != expectedInconsistentSource {
-						return fmt.Errorf("expected that node %s would report %s, instead it reports %s", workerNodes[0].Name, expectedInconsistentSource, inconsistentSources)
-					}
-
-					// Since all the other nodes consistently fail, there should also be a common result
-					mostCommonState := rootLoginCheck.Annotations[compv1alpha1.ComplianceCheckResultMostCommonAnnotation]
-					if mostCommonState != string(compv1alpha1.CheckResultFail) {
-						return fmt.Errorf("expected that there would be a common FAIL state, instead got %s", mostCommonState)
-					}
-				} else if len(workerNodes) == 2 {
-					// example: ip-10-0-184-135.us-west-1.compute.internal:PASS,ip-10-0-226-48.us-west-1.compute.internal:FAIL
-					var expectedInconsistentSource [2]string
-					expectedInconsistentSource[0] = workerNodes[0].Name + ":" + string(compv1alpha1.CheckResultPass) + "," + workerNodes[1].Name + ":" + string(compv1alpha1.CheckResultFail)
-					expectedInconsistentSource[1] = workerNodes[1].Name + ":" + string(compv1alpha1.CheckResultFail) + "," + workerNodes[0].Name + ":" + string(compv1alpha1.CheckResultPass)
-
-					inconsistentSources := rootLoginCheck.Annotations[compv1alpha1.ComplianceCheckResultInconsistentSourceAnnotation]
-					if inconsistentSources != expectedInconsistentSource[0] && inconsistentSources != expectedInconsistentSource[1] {
-						return fmt.Errorf(
-							"expected that node %s would report %s or %s, instead it reports %s",
-							workerNodes[0].Name,
-							expectedInconsistentSource[0], expectedInconsistentSource[1],
-							inconsistentSources)
-					}
-
-					// If there are only two worker nodes, we won't be able to find the common status, so both
-					// nodes would be listed as inconsistent -- we can't figure out which of the two results is
-					// consistent and which is not. Therefore this branch skips the check for
-					// compv1alpha1.ComplianceCheckResultMostCommonAnnotation
-				} else {
-					E2ELog(t, "Only one worker node? Shortcutting the test")
-					return nil
-				}
-
-				// Since all states were either pass or fail, we still create the remediation
-				workerRemediations := []string{
-					fmt.Sprintf("%s-no-direct-root-logins", workerScanName),
-				}
-				err = assertHasRemediations(t, f, suiteName, workerScanName, "worker", workerRemediations)
-				if err != nil {
-					return err
-				}
-
-				return nil
-			},
-		},
-		testExecution{
 			Name:       "TestPlatformAndNodeSuiteScan",
 			IsParallel: false,
 			TestFn: func(t *testing.T, f *framework.Framework, ctx *framework.TestCtx, namespace string) error {
